efivario/windows: fix abi confusion in variable name output
Initially it was assumed that the data returned by the
NtEnumerateSystemEnvironmentValuesEx API describes the length of
each individual entry. This assumption is wrong though since the
last entry will have the field set to 0. So it must mean that this
field describes the offset to the next entry instead and leaves the
field zero for the last entry since there is no next entry left.